// app/(main)/nyuko-list/nyuko-detail/[juchu_head_id]/[nyushuko_basho_id]/[nyushuko_dat]/_lib/types.ts
export type NyukoDetailTableValues = {
  juchuHeadId: number;
  juchuKizaiHeadIdv: number[] | null;
  juchuKizaiHeadKbnv: number;
  headNamv: string | null;
  kizaiId: number | null;
  kizaiNam: string | null;
  koenNam: string | null;
  koenbashoNam: string | null;
  kokyakuNam: string | null;
  nyushukoBashoId: number;
  nyushukoDat: string;
  nyushukoShubetuId: number | null;
  planQty: number | null;
  resultAdjQty: number | null;
  resultQty: number | null;
  sagyoKbnId: number | null;
  diff: number;
  ctnFlg: number | null;
};